+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# release-25 (2024-03-18)
+
+### Bug Fixes
+
+- resource type predicate ([#1454](https://github.com/coveo/cli/issues/1454)) ([caef795](https://github.com/coveo/cli/commits/caef795896402696b7eca6295c5cfccd1a504d82)), closes [#1453](https://github.com/coveo/cli/issues/1453)
+
+# release-24 (2024-03-14)
+
+### Bug Fixes
+
+- **snapshot:** define resourceToPreview from resources changed. ([#1453](https://github.com/coveo/cli/issues/1453)) ([c4095f8](https://github.com/coveo/cli/commits/c4095f88c917dc79cf6600c0cd897fb30713c647))
+
 # release-23 (2024-02-21)
 
 # release-22 (2024-02-21)